Updates for issue #562
Adding rule to align signal and variable assignment operators.

  1)  Updated documentation
  2)  Updated tests
      a)  Removed deprecated tests
      b)  Added process tests
  3)  Updated rules
      a)  Deprecated signal_005
      b)  Deprecated variable_assignment_005
      c)  Added rule process_400
