+++ docs/sequential_rules.rst
@@ -94,32 +94,7 @@ This rule checks the alignment of multiline sequential statements.
 sequential_005
 ##############
 
-|phase_5| |error|
-
-This rule checks the alignment of the **<=** operators over consecutive sequential lines.
-
-Following extra configurations are supported:
-
-* :code:`if_control_statements_ends_group`,
-* :code:`case_control_statements_ends_group`.
-* :code:`case_keyword_statements_ends_group`.
-* :code:`loop_control_statements_ends_group`,
-
-Refer to the section `Configuring Keyword Alignment Rules <configuring.html#configuring-keyword-alignment-rules>`_ for information on changing the configurations.
-
-**Violation**
-
-.. code-block:: vhdl
-
-   wr_en <= '1';
-   rd_en   <= '0';
-
-**Fix**
-
-.. code-block:: vhdl
-
-   wr_en <= '1';
-   rd_en <= '0';
+This rule has been deprecated and replaced with rule `process_400 <process_rules.html#process-400>`_.
 
 sequential_006
 ##############
