+++ docs/variable_assignment_rules.rst
@@ -98,31 +98,7 @@ This rule checks the alignment of multiline variable assignments.
 variable_assignment_005
 #######################
 
-|phase_5| |error|
-
-This rule checks the alignment of **:=** operators over multiple lines.
-
-Following extra configurations are supported:
-
-* :code:`if_control_statements_end_group`.
-* :code:`case_control_statements_end_group`,
-
-
-Refer to the section `Configuring Keyword Alignment Rules <configuring.html#configuring-keyword-alignment-rules>`_ for information on changing the configurations.
-
-**Violation**
-
-.. code-block:: vhdl
-
-     counter := 0;
-     count := counter + 1;
-
-**Fix**
-
-.. code-block:: vhdl
-
-     counter := 0;
-     count   := counter + 1;
+This rule has been deprecated and replaced with rule `process_400 <process_rules.html#process-400>`_.
 
 variable_assignment_006
 #######################
