+++ vsg/rules/process/__init__.py
@@ -36,4 +36,6 @@ from .rule_034 import rule_034
 from .rule_035 import rule_035
 from .rule_036 import rule_036
 
+from .rule_400 import rule_400
+
 from .rule_600 import rule_600
