+++ vsg/rules/sequential/rule_005.py
@@ -1,28 +1,9 @@
 
-from vsg.rules import align_tokens_in_region_between_tokens_unless_between_tokens
-
-from vsg import token
-
-lAlign = []
-lAlign.append(token.simple_waveform_assignment.assignment)
-lAlign.append(token.simple_force_assignment.assignment)
-lAlign.append(token.simple_release_assignment.assignment)
-
-oStart = token.process_statement.begin_keyword
-oEnd = token.process_statement.end_keyword
-
-lUnless = []
+from vsg.depricated_rule import Depricated
 
 
-class rule_005(align_tokens_in_region_between_tokens_unless_between_tokens):
-    '''
-    Ensures the alignment of the "<=" keyword over multiple lines.
-    '''
+class rule_005(Depricated):
 
     def __init__(self):
-        align_tokens_in_region_between_tokens_unless_between_tokens.__init__(self, 'sequential', '005', lAlign, oStart, oEnd, lUnless)
-        self.solution = 'Align identifer.'
-        self.if_control_statements_ends_group = True
-        self.case_control_statements_ends_group = True
-        self.case_keyword_statements_ends_group = False
-        self.loop_control_statements_ends_group = False
+        Depricated.__init__(self, 'sequential', '005')
+        self.message.append('Rule ' + self.unique_id + ' has been merged into process_400.')
